Improved merge strategy for en content (#182)

+++ src/beeware_docs_tools/build_md_translations.py
@@ -167,6 +167,27 @@ def main():
                 build_with_warnings=args.build_with_warnings,
             )
 
+            # For every non-english language output: if there are language-folder
+            # redirects, remove those from the translated version.
+            if language != "en":
+                for redirect_lang in args.language_code:
+                    lang_redirect = output / language / redirect_lang
+                    if lang_redirect.is_dir():
+                        print("Prune", lang_redirect)
+                        shutil.rmtree(lang_redirect)
+                    lang_redirect = (
+                        output / language / redirect_lang.replace("_", "-").lower()
+                    )
+                    if lang_redirect.is_dir():
+                        print("Prune", lang_redirect)
+                        shutil.rmtree(lang_redirect)
+
+            # If a SUMMARY/index.md has been created, prune it as well.
+            summary = output / language / "SUMMARY"
+            if summary.is_dir():
+                print("Prune", summary)
+                shutil.rmtree(summary)
+
         # If we've built EN, move the content into the root. The `en`` folder
         # may contain an `en` folder; allow for that edge case by moving `en`
         # to a safe location first.
@@ -174,8 +195,14 @@ def main():
             en_orig = output / "en-nested"
             shutil.move(output / "en", en_orig)
             for path in en_orig.iterdir():
-                shutil.rmtree(output / path.name, ignore_errors=True)
-                shutil.move(path, output / path.name)
+                if (output / path.name).is_dir():
+                    print("Overlay ", path, output / path.name)
+                    shutil.copytree(path, output / path.name, dirs_exist_ok=True)
+                else:
+                    print("Move", path, output / path.name)
+                    shutil.rmtree(output / path.name, ignore_errors=True)
+                    shutil.move(path, output / path.name)
+            print("Prune", en_orig)
             shutil.rmtree(en_orig)
